requirements.txt
discord.py
flask
python-dotenv
fastapi==0.68.0
uvicorn==0.15.0